backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.1
APScheduler==3.10.4
PyJWT==2.8.0
qrcode[pil]==7.4.2
waitress==3.0.0

